Add DocumentVoter, relation between User & Document

+++ src/Controller/DocumentController.php
@@ -29,6 +29,7 @@ class DocumentController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
+            $document->setUser($this->getUser());
             $entityManager->persist($document);
             $entityManager->flush();
 
@@ -67,40 +68,12 @@ class DocumentController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="document_show", methods={"GET"})
-     */
-    public function show(Document $document): Response
-    {
-        return $this->render('document/show.html.twig', [
-            'document' => $document,
-        ]);
-    }
-
-    /**
-     * @Route("/{id}/edit", name="document_edit", methods={"GET","POST"})
-     */
-    public function edit(Request $request, Document $document): Response
-    {
-        $form = $this->createForm(DocumentType::class, $document);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('document_index');
-        }
-
-        return $this->render('document/edit.html.twig', [
-            'document' => $document,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
      * @Route("/{id}", name="document_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Document $document): Response
     {
+        $this->denyAccessUnlessGranted('DELETE', $document);
+
         if ($this->isCsrfTokenValid('delete'.$document->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($document);
